// src/openapi/models/document-space-app-client-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DocumentSpaceAppClientResponseDto
 */
export interface DocumentSpaceAppClientResponseDto {
    /**
     * 
     * @type {string}
     * @memberof DocumentSpaceAppClientResponseDto
     */
    appClientId: string;
    /**
     * 
     * @type {string}
     * @memberof DocumentSpaceAppClientResponseDto
     */
    appClientName: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof DocumentSpaceAppClientResponseDto
     */
    privileges: Array<DocumentSpaceAppClientResponseDtoPrivilegesEnum>;
}

/**
    * @export
    * @enum {string}
    */
export enum DocumentSpaceAppClientResponseDtoPrivilegesEnum {
    Read = 'READ',
    Write = 'WRITE',
    Membership = 'MEMBERSHIP'
}